+++ EducacaoXpert/src/EducacaoXpert.PagamentoFaturamento.Anticorruption/Interfaces/IPayPalGateway.cs
@@ -6,5 +6,5 @@ public interface IPayPalGateway
 {
     string GetPayPalServiceKey(string apiKey, string encriptionKey);
     string GetCardHashKey(string serviceKey, string cartaoCredito);
-    Transacao CommitTransaction(string cardHashKey, string orderId, decimal amount);
+    Transacao CommitTransaction(string cardHashKey, string orderId, int amount);
 }
